+++ Kaats-collection/collectionlist.js
@@ -1,0 +1,95 @@
+
+const collection = [
+    {
+      title: 'Serve The Servants',
+      artist: 'Nirvana',
+      album: 'In Utero',
+      genre: 'Grunge',
+      picture: 'images/pic1.png',
+      spotify: 'https://open.spotify.com/track/3w5Ekq9O8g0gGFEsT77Ydi?si=864a486db7f84e4a'
+    },
+
+    {
+      title: 'The Pot',
+      artist: 'Tool',
+      album: '10 000 Days',
+      genre: 'Rock',
+      picture: 'images/pic2.png',
+      spotify: 'https://open.spotify.com/track/1lATXTBJDHwawvT1UfxWu3?si=dbb3d83742184f97'
+    },  
+
+    {
+      title: 'Are you in',
+      artist: 'Incubus',
+      album: 'Morning view',
+      genre: 'Rock',
+      picture: 'images/pic3.png',
+      spotify: 'https://open.spotify.com/track/1MvJno497VkQR3RsiJcRVm?si=44ea993abb294425'
+    },
+
+    {
+      title: 'Go with the Flow',
+      artist: 'Queens of the Stoneage',
+      album: 'Songs for the Deaf',
+      genre: 'Rock',
+      picture: 'images/pic4.png',
+      spotify: 'https://open.spotify.com/track/0wzRcekWyVCSyPtlPOeJau?si=d9456862cae345a2'
+    },
+
+    {
+      title: 'Pardon me',
+      artist: 'Incubus',
+      album: 'Make Yourself',
+      genre: 'Rock',
+      picture: 'images/pic3.png',
+      spotify: 'https://open.spotify.com/track/1Y13csEpu3TK5gQdzGLrd8?si=0d8402ad75fb4794'
+    },
+
+    {
+      title: 'Everlong',
+      artist: 'Foo Fighters',
+      album: 'The Colour And The Shape',
+      genre: 'Rock',
+      picture: 'images/pic6.png',
+      spotify: 'https://open.spotify.com/track/5UWwZ5lm5PKu6eKsHAGxOk?si=846e5f498e344238'
+    },
+
+    {
+      title: 'Rocketman',
+      artist: 'Elton John',
+      album: 'Honky Chateau',
+      genre: 'Pop',
+      picture: 'images/pic7.png',
+      spotify: 'https://open.spotify.com/track/3gdewACMIVMEWVbyb8O9sY?si=461a8be840034050'
+    }, 
+
+    {
+      title: 'As',
+      artist: 'Stevie Wonder',
+      album: 'Songs In The Key Of Life',
+      genre: 'Soul',
+      picture: 'images/pic8.png',
+      spotify: 'https://open.spotify.com/track/3hFh4tsMFhskEXlJ5DCzsG?si=a9b2388a13634c96'
+    }, 
+
+    {
+      title: 'Are You Mine',
+      artist: 'Arctic Monkeys',
+      album: 'AM',
+      genre: 'Rock',
+      picture: 'images/pic9.png',
+      spotify: 'https://open.spotify.com/track/2AT8iROs4FQueDv2c8q2KE?si=110d2f2324634b70'
+    },  
+
+    {
+      title: 'Give It All',
+      artist: 'Rise Against',
+      album: 'Siren Song Of The Counter-Culture',
+      genre: 'Punk Rock',
+      picture: 'images/pic10.png',
+      spotify: 'https://open.spotify.com/track/5O6SumCYk8Ug81UbfXAPJC?si=14102ce2d44140dd'
+    }
+  ]
+
+
+export default collection
