+++ Kaats-collection/index-v1.html
@@ -21,6 +21,6 @@
     </section>
    
 
-<script src="collection-v1.js"></script>
+<script type="module" src="collection-v1.js"></script>
 </body>
 </html>
